Criacao do metodo get em BoletosManager
