+++ src/main/java/br/com/pjbank/sdk/enums/StatusPagamentoBoleto.java
@@ -5,13 +5,13 @@ package br.com.pjbank.sdk.enums;
  * @author Paulo Vitor Braga Pessoa <foxpv213@gmail.com>
  *
  */
-public enum Pago {
+public enum StatusPagamentoBoleto {
 	PAGO("1"),
 	ABERTO("0");
 
     private String name;
 
-    Pago(String name) {
+    StatusPagamentoBoleto(String name) {
         this.name = name;
     }
 
@@ -19,8 +19,8 @@ public enum Pago {
         return this.name;
     }
 
-    public static Pago fromString(String name) {
-        for (Pago obj : Pago.values()) {
+    public static StatusPagamentoBoleto fromString(String name) {
+        for (StatusPagamentoBoleto obj : StatusPagamentoBoleto.values()) {
             if (obj.name.equalsIgnoreCase(name)) {
                 return obj;
             }
